removing _is_saved_in_carto 1
