+++ test/mocks/dataset_mock.py
@@ -29,6 +29,15 @@ class DataFrameDatasetMock(DataFrameDataset):
     def exists(self):
         return False
 
+    def _create_table(self, _, _2):
+        return True
+
+    def _copyfrom(self, _, _2):
+        return True
+
+    def compute_geom_type(self):
+        return Dataset.GEOM_TYPE_POINT
+
 
 class QueryDatasetMock(QueryDataset):
     def _create_client(self):
@@ -38,6 +47,13 @@ class QueryDatasetMock(QueryDataset):
         return False
 
     def _create_table_from_query(self):
+        return True
+
+    def _copyto(self, _, _2, _3, _4, _5):
+        self._df = pd.DataFrame({'column_name': [1]})
+        return self._df
+
+    def _get_query_columns(self):
         return True
 
 
@@ -50,6 +66,16 @@ class TableDatasetMock(TableDataset):
 
     def _get_dataset_info(self):
         return DatasetInfoMock(self._context, self._table_name)
+
+    def _copyto(self, _, _2, _3, _4, _5):
+        self._df = pd.DataFrame({'column_name': [1]})
+        return self._df
+
+    def _get_table_columns(self):
+        return True
+
+    def _get_read_query(self, _, _2):
+        return True
 
 
 class DatasetMock(Dataset):
@@ -71,16 +97,3 @@ class DatasetMock(Dataset):
             strategy = QueryDatasetMock
 
         super(DatasetMock, self)._set_strategy(strategy, data, context, schema)
-
-    def download(self):
-        self._df = pd.DataFrame({'column_name': [1]})
-        return self._df
-
-    def _copyfrom(self, _):
-        return True
-
-    def _create_table(self, _):
-        return True
-
-    def compute_geom_type(self):
-        return Dataset.GEOM_TYPE_POINT
